chores: make it the component for Toast much cleaner
