+++ .vscode/settings.json
@@ -1,0 +1,5 @@
+{
+    "eslint.format.enable": true,
+    "eslint.packageManager": "yarn",
+    "eslint.run": "onSave"
+}
