+++ common/types.ts
@@ -24,6 +24,6 @@ export interface ToastMessage {
 
 export interface SubmitForm {
     inputTextRef: React.Ref<HTMLInputElement | HTMLElement | string | any>;
-    handleSubmit: any;
+    handleSubmit?: any;
     loadingMutation: boolean;
 }
