+++ components/events/Toast.tsx
@@ -1,27 +1,22 @@
 import { ToastMessage } from "common/types";
 
 export default function Toast (toastMessage: ToastMessage) {
-  return (
-   <>
-     {!!toastMessage.message && (
-        <div className="toast toast-top toast-end">
-          <div className="alert alert-success">
-            <div>
-              <span>{toastMessage.message}</span>
-            </div>
-          </div>
-        </div>
-      )}
+  const messageType = () => {
+    if(!!toastMessage.message) {
+      return 'success';
+    }
+    if(!!toastMessage.error) {
+      return 'error';
+    }
+  };
 
-      {!!toastMessage.error && (
-        <div className="toast toast-top toast-end">
-          <div className="alert alert-error">
-            <div>
-              <span>{toastMessage.error}</span>
-            </div>
-          </div>
+  return (
+    <div className="toast toast-top toast-end">
+      <div className={`alert alert-${messageType}`}>
+        <div>
+          <span>{!!toastMessage.message ? toastMessage.message : toastMessage.error}</span>
         </div>
-      )}
-   </>
+      </div>
+    </div>
   )
 }
